requirements.txt
requests>=2.25.1
moviepy>=1.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.4.0
google-auth-httplib2>=0.1.0
aiohttp>=3.8.0  # For asynchronous HTTP requests
tenacity>=8.0.0  # For retry logic